requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
tifffile
scikit-learn
tqdm
matplotlib